Move code to generate bundled plugconf to plugconf.go
